requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
alembic==1.13.1
asyncpg==0.29.0
prisma==0.11.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2
pyotp==2.9.0
qrcode[pil]==7.4.2
cryptography>=41.0.0,<42.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
websockets==12.0
aiofiles==23.2.1
Pillow==10.1.0
requests==2.31.0
schedule==1.2.1
geoip2==5.1.0
maxminddb-geolite2==2018.703
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0 